backend_problem_handler/requirements.txt
Django==5.2.6
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
dotenv==0.9.9
python-dotenv==1.1.1
pydantic==2.11.7
pyyaml==6.0.2
natsort==8.4.0
pika==1.3.2
psycopg2-binary==2.9.10
django-cors-headers==4.8.0